feat: update admin product routes

+++ routes/web.php
@@ -99,12 +99,12 @@ Route::group(['middleware' => 'auth'], function () {
 		Route::get('/', [HomeController::class, 'index'])->name('admin.home');
 		
 		Route::prefix('/products')->group(function () {
-			Route::get('/', [AdminProductController::class, 'index'])->name('products.index');
-			Route::get('/create', [AdminProductController::class, 'create'])->name('products.create');
-			Route::post('/', [AdminProductController::class, 'store'])->name('products.store');
-			Route::get('/edit/{id}', [AdminProductController::class, 'edit'])->name('products.edit');
-			Route::patch('/{id}', [AdminProductController::class, 'update'])->name('products.update');
-			Route::delete('/{id}', [AdminProductController::class, 'destroy'])->name('products.destroy');
+			Route::get('/', [AdminProductController::class, 'index'])->name('admin.products.index');
+			Route::get('/create', [AdminProductController::class, 'create'])->name('admin.products.create');
+			Route::post('/', [AdminProductController::class, 'store'])->name('admin.products.store');
+			Route::get('/edit/{id}', [AdminProductController::class, 'edit'])->name('admin.products.edit');
+			Route::patch('/{id}', [AdminProductController::class, 'update'])->name('admin.products.update');
+			Route::delete('/{id}', [AdminProductController::class, 'destroy'])->name('admin.products.destroy');
 		});
 
 		Route::prefix('/users')->group(function () {
